+++ autometrics/src/lib.rs
@@ -212,11 +212,13 @@ pub mod __private {
     #[cfg(debug_assertions)]
     use crate::objectives::Objective;
     use crate::task_local::LocalKey;
+    use once_cell::sync::OnceCell;
     use std::{cell::RefCell, thread_local};
 
     pub use crate::constants::*;
     pub use crate::labels::*;
     pub use crate::tracker::{AutometricsTracker, TrackMetrics};
+    pub use spez::spez;
 
     /// Task-local value used for tracking which function called the current function
     pub static CALLER: LocalKey<&'static str> = {
@@ -230,8 +232,6 @@ pub mod __private {
 
         LocalKey { inner: CALLER_KEY }
     };
-
-    pub use spez::spez;
 
     // Re-export linkme so that it can be used by the macro-generated code
     #[cfg(debug_assertions)]
@@ -252,6 +252,7 @@ pub mod __private {
         pub name: &'static str,
         pub module: &'static str,
         pub objective: Option<Objective>,
+        pub service_name: &'static str,
     }
 
     #[cfg(debug_assertions)]
@@ -268,6 +269,7 @@ pub mod __private {
             CounterLabels {
                 function: function.name,
                 module: function.module,
+                service_name: function.service_name,
                 caller: "",
                 result: Some(ResultLabel::Ok),
                 ok: None,
@@ -276,5 +278,21 @@ pub mod __private {
                 objective_percentile,
             }
         }
+    }
+
+    /// Load the service name from one of the following runtime environment variables:
+    /// - `AUTOMETRICS_SERVICE_NAME`
+    /// - `OTEL_SERVICE_NAME`
+    /// Or, fall back to the name of the cargo package defined in the `Cargo.toml` file.
+    pub fn service_name(cargo_pkg_name: &'static str) -> &'static str {
+        // Note that we are using a OnceCell here because we want the label value to
+        // be available as a &'static str. This allows us to load the value from the
+        // environment variables once at startup while still accessing it as a &'static str
+        static SERVICE_NAME: OnceCell<String> = OnceCell::new();
+        SERVICE_NAME.get_or_init(|| {
+            std::env::var("AUTOMETRICS_SERVICE_NAME")
+                .or_else(|_| std::env::var("OTEL_SERVICE_NAME"))
+                .unwrap_or_else(|_| cargo_pkg_name.to_string())
+        })
     }
 }