+++ autometrics/src/tracker/metrics.rs
@@ -5,7 +5,7 @@ use crate::labels::{BuildInfoLabels, CounterLabels, GaugeLabels, HistogramLabels
 use crate::tracker::TrackMetrics;
 use metrics::{
     describe_counter, describe_gauge, describe_histogram, register_counter, register_gauge,
-    register_histogram, Gauge,
+    register_histogram, Gauge, Unit,
 };
 use std::{sync::Once, time::Instant};
 
@@ -14,9 +14,13 @@ static SET_BUILD_INFO: Once = Once::new();
 
 fn describe_metrics() {
     DESCRIBE_METRICS.call_once(|| {
-        describe_counter!(COUNTER_NAME, COUNTER_DESCRIPTION);
-        describe_histogram!(HISTOGRAM_NAME, HISTOGRAM_DESCRIPTION);
-        describe_gauge!(GAUGE_NAME, GAUGE_DESCRIPTION);
+        describe_counter!(COUNTER_NAME_PROMETHEUS, COUNTER_DESCRIPTION);
+        describe_histogram!(
+            HISTOGRAM_NAME_PROMETHEUS,
+            Unit::Seconds,
+            HISTOGRAM_DESCRIPTION
+        );
+        describe_gauge!(GAUGE_NAME_PROMETHEUS, GAUGE_DESCRIPTION);
         describe_gauge!(BUILD_INFO_NAME, BUILD_INFO_DESCRIPTION);
     });
 }
@@ -46,8 +50,8 @@ impl TrackMetrics for MetricsTracker {
 
     fn finish(self, counter_labels: &CounterLabels, histogram_labels: &HistogramLabels) {
         let duration = self.start.elapsed().as_secs_f64();
-        register_counter!(COUNTER_NAME, &counter_labels.to_vec()).increment(1);
-        register_histogram!(HISTOGRAM_NAME, &histogram_labels.to_vec()).record(duration);
+        register_counter!(COUNTER_NAME_PROMETHEUS, &counter_labels.to_vec()).increment(1);
+        register_histogram!(HISTOGRAM_NAME_PROMETHEUS, &histogram_labels.to_vec()).record(duration);
         if let Some(gauge) = self.gauge {
             gauge.decrement(1.0);
         }