+++ CHANGELOG.md
@@ -11,14 +11,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Basic benchmarking
+- All metrics now have a `service.name` label attached. This is set via runtime environment
+  variable (`AUTOMETRICS_SERVICE_NAME` or `OTEL_SERVICE_NAME`), or falls back to the crate
+  name defined in the `Cargo.toml`
 - Function counters are initialized to zero in debug builds. This exposes details of
   instrumented functions to Prometheus and visualization tools built on top of it,
   before the functions have been called.
+- Basic benchmarking
 
 ### Changed
 
--
+- Renamed the `function.calls.count` metric to `function.calls` (which is exported
+  to Prometheus as `function_calls_total`) to be in line with OpenTelemetry and
+  OpenMetrics naming conventions. **Dashboards and alerting rules must be updated.**
+- When the `function.calls.duration` histogram is exported to Prometheus, it now
+  includes the units (`function_calls_duration_seconds`) to be in line with
+  Prometheus/OpenMetrics naming conventions. **Dashboards and alerting rules must be updated.**
 
 ### Deprecated